SlackAPI/requirements.txt
requests
python-dotenv
slack-sdk
fastapi
uvicorn
